+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   `clone` and `extend` are used in all examples unfortunately so the community must have adopted them.
   Devs using `extend` on classes should mutate the prototype directly (or with `defineProperty`) or subclass.
   Using `clone` or `extend` to assign to an object was always a bad idea. Use lodash or whatever.
+- chore(TS): Followup for interactivy and controls migration to TS [#8404](https://github.com/fabricjs/fabric.js/pull/8404)
+- refactor(IText): Fixes Draggable Text for retina and viewport transform #8534
+- chore(TS): refactor canvas init, fix `_initRetinaScaling` regression #8520
 - chore(TS): remove all remaining empty declarations [#8593](https://github.com/fabricjs/fabric.js/pull/8593)
 - refactor(IText): modernize IText cursor animation based on animation API changes (and fix minor regression) plus leftovers from #8547 [#8583](https://github.com/fabricjs/fabric.js/pull/8583)
 - refactor(Canvas, IText): Handle cross instance text editing states to an EditingManager class [#8543](https://github.com/fabricjs/fabric.js/pull/8543)